requirements.txt
google-auth
google-api-python-client
openai
fastapi
uvicorn[standard]
pandas
sqlalchemy
python-dotenv
google-auth-oauthlib
apscheduler
jinja2
numpy 